feat(open-spark): implemented landing page 1. implemented reusable carousel component for landing page 2. implemented reusable button with prefix and postfix icon with text
